move filepath to kak package
