+++ kak/filepath.go
@@ -1,4 +1,4 @@
-package cmd
+package kak
 
 import (
 	"os"
