+++ kak/send.go
@@ -6,7 +6,7 @@ import (
 	"os/exec"
 )
 
-func Send(kctx Context, command string) error {
+func Send(kctx *Context, command string) error {
 	cmd := exec.Command("kak", "-p", kctx.Session.Name)
 	// cmd.Stdout = os.Stdout
 	// cmd.Stderr = os.Stderr
